requirements.txt
altair==4.1.0
argcomplete==1.12.3
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.0.0
blinker==1.4
blis==0.2.4
boto3==1.18.20
botocore==1.21.20
cachetools==4.2.2
catalogue==2.0.5
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
configparser==5.0.2
cycler==0.10.0
cymem==2.0.5
datasets==1.11.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
docker-pycreds==0.4.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.1.0/en_core_web_lg-2.1.0.tar.gz
entrypoints==0.3
filelock==3.0.12
fsspec==2021.7.0
gensim==3.8.3
gitdb==4.0.7
GitPython==3.1.18
huggingface-hub==0.0.12
idna==3.2
importlib-metadata==4.6.3
ipykernel==6.1.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==2.6.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
multiprocess==0.70.12.2
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
neuralcoref==4.0
nltk==3.6.2
notebook==6.4.3
numpy==1.21.1
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pkg_resources==0.0.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==5.0.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.3
requests==2.26.0
s3transfer==0.5.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
sentence-transformers==2.0.0
sentencepiece==0.1.96
sentry-sdk==1.3.1
seqeval==1.2.2
shortuuid==1.0.1
simpletransformers==0.61.13
six==1.16.0
smart-open==5.1.0
smmap==4.0.0
spacy==2.1.1
srsly==1.0.5
stanford-openie==1.1.0
stanfordnlp==0.2.0
streamlit==0.86.0
subprocess32==3.5.4
tensorboardX==2.4
terminado==0.11.0
testpath==0.5.0
thinc==7.0.8
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
transformers==4.9.2
typer==0.3.2
typing-extensions==3.10.0.0
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
wandb==0.12.0
wasabi==0.8.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
wget==3.2
widgetsnbextension==3.5.1
xxhash==2.0.2
zipp==3.5.0
